remove backtrace from optee path.
