+++ enclave/core/optee/backtrace.c
@@ -11,3 +11,16 @@ int oe_backtrace(void** buffer, int size)
 
     return 0;
 }
+
+char** oe_backtrace_symbols(void* const* buffer, int size)
+{
+    OE_UNUSED(buffer);
+    OE_UNUSED(size);
+
+    return NULL;
+}
+
+void oe_backtrace_symbols_free(char** ptr)
+{
+    OE_UNUSED(ptr);
+}
